+++ epochStart/bootstrap/process_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/multiversx/mx-chain-go/epochStart/mock"
 	errorsMx "github.com/multiversx/mx-chain-go/errors"
 	"github.com/multiversx/mx-chain-go/process"
+	processMocks "github.com/multiversx/mx-chain-go/process/mock"
 	"github.com/multiversx/mx-chain-go/sharding"
 	"github.com/multiversx/mx-chain-go/sharding/nodesCoordinator"
 	"github.com/multiversx/mx-chain-go/state"
@@ -243,7 +244,7 @@ func createMockEpochStartBootstrapArgs(
 		},
 		TrieSyncStatisticsProvider: &testscommon.SizeSyncStatisticsHandlerStub{},
 		StateStatsHandler:          disabledStatistics.NewStateStatistics(),
-		RunTypeComponents:          mock.NewRunTypeComponentsStub(),
+		RunTypeComponents:          processMocks.NewRunTypeComponentsStub(),
 	}
 }
 
@@ -651,7 +652,7 @@ func TestNewEpochStartBootstrap_NilArgsChecks(t *testing.T) {
 
 		coreComp, cryptoComp := createComponentsForEpochStart()
 		args := createMockEpochStartBootstrapArgs(coreComp, cryptoComp)
-		rtMock := mock.NewRunTypeComponentsStub()
+		rtMock := processMocks.NewRunTypeComponentsStub()
 		rtMock.ShardCoordinatorFactory = nil
 		args.RunTypeComponents = rtMock
 		epochStartProvider, err := NewEpochStartBootstrap(args)
@@ -663,7 +664,7 @@ func TestNewEpochStartBootstrap_NilArgsChecks(t *testing.T) {
 
 		coreComp, cryptoComp := createComponentsForEpochStart()
 		args := createMockEpochStartBootstrapArgs(coreComp, cryptoComp)
-		rtMock := mock.NewRunTypeComponentsStub()
+		rtMock := processMocks.NewRunTypeComponentsStub()
 		rtMock.AdditionalStorageServiceFactory = nil
 		args.RunTypeComponents = rtMock
 		epochStartProvider, err := NewEpochStartBootstrap(args)
@@ -675,7 +676,7 @@ func TestNewEpochStartBootstrap_NilArgsChecks(t *testing.T) {
 
 		coreComp, cryptoComp := createComponentsForEpochStart()
 		args := createMockEpochStartBootstrapArgs(coreComp, cryptoComp)
-		rtMock := mock.NewRunTypeComponentsStub()
+		rtMock := processMocks.NewRunTypeComponentsStub()
 		rtMock.NodesCoordinatorWithRaterFactory = nil
 		args.RunTypeComponents = rtMock
 		epochStartProvider, err := NewEpochStartBootstrap(args)
