--- epochStart/mock/runTypeComponentsStub.go
@@ -1,44 +1,0 @@
-package mock
-
-import (
-	"github.com/multiversx/mx-chain-go/process"
-	"github.com/multiversx/mx-chain-go/sharding"
-	"github.com/multiversx/mx-chain-go/sharding/nodesCoordinator"
-	"github.com/multiversx/mx-chain-go/testscommon"
-)
-
-// RunTypeComponentsStub -
-type RunTypeComponentsStub struct {
-	AdditionalStorageServiceFactory  process.AdditionalStorageServiceCreator
-	ShardCoordinatorFactory          sharding.ShardCoordinatorFactory
-	NodesCoordinatorWithRaterFactory nodesCoordinator.NodesCoordinatorWithRaterFactory
-}
-
-// NewRunTypeComponentsStub -
-func NewRunTypeComponentsStub() *RunTypeComponentsStub {
-	return &RunTypeComponentsStub{
-		AdditionalStorageServiceFactory:  &testscommon.AdditionalStorageServiceFactoryMock{},
-		ShardCoordinatorFactory:          sharding.NewMultiShardCoordinatorFactory(),
-		NodesCoordinatorWithRaterFactory: nodesCoordinator.NewIndexHashedNodesCoordinatorWithRaterFactory(),
-	}
-}
-
-// AdditionalStorageServiceCreator -
-func (r *RunTypeComponentsStub) AdditionalStorageServiceCreator() process.AdditionalStorageServiceCreator {
-	return r.AdditionalStorageServiceFactory
-}
-
-// ShardCoordinatorCreator -
-func (r *RunTypeComponentsStub) ShardCoordinatorCreator() sharding.ShardCoordinatorFactory {
-	return r.ShardCoordinatorFactory
-}
-
-// NodesCoordinatorWithRaterCreator -
-func (r *RunTypeComponentsStub) NodesCoordinatorWithRaterCreator() nodesCoordinator.NodesCoordinatorWithRaterFactory {
-	return r.NodesCoordinatorWithRaterFactory
-}
-
-// IsInterfaceNil -
-func (r *RunTypeComponentsStub) IsInterfaceNil() bool {
-	return r == nil
-}
